refactor(map): map将action抽象到独立文件
